requirements.txt
Flask==1.1.1
pytest
pylint
pandas
google-cloud-language
wikipedia